requirements.txt
pandas
psycopg2-binary
streamlit
pytest
pylint
python-dotenv
plotly